Extract add item form into separate component and move to List view
