+++ src/components/index.js
@@ -1,2 +1,3 @@
+export * from './AddItem';
 export * from './ListItem';
 export * from './SingleList';
